+++ setup/constant.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+SSL_CERTIFICATE=fullchain.pem
+SSL_CERTIFICATE_KEY=privkey.pem